trying this without using async methods, as there is no UI to interupt. even found a way to still use the download progress and download complete events that are only for async methods

+++ ReditScraper/Program.cs
@@ -6,16 +6,16 @@ using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Threading.Tasks;
+using System.Threading;
 
 namespace RedditScraper
 {
 	class Program
 	{
 		private static string inputSubreddit;
+		private static Subreddit subreddit;
 		private static string directory;
 		private static FromTime time;
-		private static WebClient wc;
 		private static int amount;
 
 		public static void Main()
@@ -23,12 +23,14 @@ namespace RedditScraper
 			if (!System.Net.NetworkInformation.NetworkInterface.GetIsNetworkAvailable()) return;
 
 			GetUserInput();
-			DownloadRedditPostsAsync();
+			DownloadRedditPosts();
+			AttemptFileCleanup();
+			Console.ReadLine();
 		}
 
 		private static void GetUserInput()
 		{
-			Console.WriteLine("Subreddit?");
+			Console.WriteLine("Subreddit? (default \"funny\")");
 			inputSubreddit = Console.ReadLine();
 			inputSubreddit = !string.IsNullOrWhiteSpace(inputSubreddit) ? inputSubreddit : "funny";
 
@@ -36,54 +38,50 @@ namespace RedditScraper
 			var inputAmount = Console.ReadLine();
 			amount = !string.IsNullOrEmpty(inputAmount) ? int.Parse(inputAmount) : 100;
 
-			Console.WriteLine("Time Period?");
+			Console.WriteLine("Time Period? (default \"All Time\")");
 			Console.WriteLine("0 = All Time");
 			Console.WriteLine("1 = Past Year");
 			Console.WriteLine("2 = Past Month");
 			Console.WriteLine("3 = Past Week");
 			Console.WriteLine("4 = Past Day");
 			Console.WriteLine("5 = Past Hour");
-			string timeInput = Console.ReadLine();
-			if (!Enum.TryParse(timeInput, out FromTime time)) time = FromTime.All;
+			if (!Enum.TryParse(Console.ReadLine(), out FromTime time)) time = FromTime.All;
 		}
 
-		private static void DownloadRedditPostsAsync()
+		private static void DownloadRedditPosts()
 		{
 			directory = $@"C:\reddit\{inputSubreddit}\";
 			Console.WriteLine($"Creating directory at \"{directory}\"");
 			Directory.CreateDirectory(directory);
 
-			var subreddit = new Reddit().GetSubreddit($"/r/{inputSubreddit}");
+			subreddit = new Reddit().GetSubreddit($"/r/{inputSubreddit}");
 			Console.WriteLine($"Looking on {subreddit} for {amount} posts...");
 
 			var foundPosts = subreddit.GetTop(time).Take(amount);
 			Console.WriteLine($"Found {foundPosts.Count()} posts on {subreddit}");
 
-			var posts = foundPosts.Select(x => x.Url.ToString());
-			DownloadImagesAsync(posts);
+			foundPosts.ToList().ForEach(x => DownloadImage(x.Url.ToString()));
 
-			Console.ReadLine();
+			var files = Directory.GetFiles(directory);
+			Console.WriteLine($"Downloaded {files.Length} files from  {subreddit}...");
 		}
 
-		private static async Task DownloadImagesAsync(IEnumerable<string> posts)
+		private static void AttemptFileCleanup()
 		{
-			await Task.WhenAll(posts.Select(i => DownloadImages(i)));
+			Console.WriteLine($"Attempting to clean up bad files...");
+			DeleteBadFiles();
+
+			var files = Directory.GetFiles(directory);
+			Console.WriteLine($"{files.Length} remain from {subreddit}... Enjoy!");
 		}
 
-		private static async Task DownloadImages(string imageURL)
+		private static void DownloadImage(string imageURL)
 		{
 			FixImageUrl(ref imageURL);
 			string fileName = CreateFileName(imageURL);
+			string path = directory + fileName;
 
-			try
-			{
-				wc = new WebClient();
-				wc.QueryString.Add("fileName", fileName);
-				wc.DownloadProgressChanged += Wc_DownloadProgressChanged;
-				wc.DownloadFileCompleted += Wc_DownloadFileCompleted;
-				await wc.DownloadFileTaskAsync(new Uri(imageURL), directory + fileName);
-			}
-			catch { }
+			DownloadFile(new Uri(imageURL), path, fileName);
 		}
 
 		private static void FixImageUrl(ref string imageUrl)
@@ -113,37 +111,56 @@ namespace RedditScraper
 			return fileName;
 		}
 
+		public static void DownloadFile(Uri uri, string destination, string fileName)
+		{
+			using (var wc = new WebClient())
+			{
+				wc.QueryString.Add("fileName", fileName);
+				wc.DownloadProgressChanged += Wc_DownloadProgressChanged;
+				wc.DownloadFileCompleted += Wc_DownloadFileCompleted;
+
+				var syncObject = new object();
+				lock (syncObject)
+				{
+					wc.DownloadFileAsync(uri, destination, syncObject);
+					Monitor.Wait(syncObject);
+				}
+			}
+		}
+
 		private static void Wc_DownloadProgressChanged(object sender, DownloadProgressChangedEventArgs e)
 		{
 			var fileName = ((WebClient)(sender)).QueryString["fileName"];
 		
-			string progress = new string('#', (int) Math.Round(e.ProgressPercentage / 5.0));
-			string progressBar = progress + "--------------------".Substring(0, 20 - progress.Length);
+			string progress = new string('#', (int) Math.Round(e.ProgressPercentage / 2.0));
+			string progressBar = progress + "--------------------------------------------------".Substring(0, 50 - progress.Length);
+			string progressPercent = "000".Substring(0, 3 - e.ProgressPercentage.ToString().Length) + e.ProgressPercentage;
 
-			Console.WriteLine($"{fileName} - {e.ProgressPercentage}% {progressBar}");
+			Console.WriteLine($"{fileName} - {progressPercent}% {progressBar}");
 		}
 
 		private static void Wc_DownloadFileCompleted(object sender, AsyncCompletedEventArgs e)
 		{
-			wc.Dispose();
-			try
+			lock (e.UserState)
 			{
-				var fileName = ((WebClient)(sender)).QueryString["fileName"];
-				var fileInfo = new FileInfo(directory + fileName);
-				if 
-				(
-					fileInfo != null &&
-					fileInfo.Length == 0 || // Broken file
-					fileInfo.Length == 503 ||  // Imgur's 'missing' image 
-					string.IsNullOrWhiteSpace(fileInfo.Extension) || // Missing extension
-					fileInfo.Extension.Length > 4 // URL parameters probably passed into filename
-				)
+				Monitor.Pulse(e.UserState);
+			}
+		}
+
+		private static void DeleteBadFiles()
+		{
+			var files = Directory.GetFiles(directory);
+			var deletedFiles = 0;
+			foreach (var file in files)
+			{
+				var fileInfo = new FileInfo(file);
+				if (fileInfo != null && fileInfo.Length == 0 || fileInfo.Length == 503 ||  string.IsNullOrWhiteSpace(fileInfo.Extension) || fileInfo.Extension.Length > 4)
 				{
 					fileInfo.Delete();
+					deletedFiles++;
 				}
 			}
-			catch (IOException) { }
-			catch (ArgumentException) { }
+			Console.WriteLine($"Deleted {deletedFiles} files...");
 		}
 	}
 }
